+++ test/cruds/removing-rows.spec.ts
@@ -12,7 +12,7 @@ import {
   expectReferenceToHaveRefError,
   extractMatrixRange,
   extractRange,
-  extractReference,
+  extractReference, noSpace,
   verifyRangesInSheet,
   verifyValues,
 } from '../testUtils'
@@ -548,7 +548,7 @@ describe('Removing rows - reevaluation', () => {
   })
 })
 
-describe('Removing rows - matrices', () => {
+describe('Removing rows - arrays', () => {
   it('should not remove row with formula matrix', () => {
     const engine = HyperFormula.buildFromArray([
       ['1', '2'],
@@ -601,6 +601,110 @@ describe('Removing rows - matrices', () => {
     engine.removeRows(0, [1, 1])
 
     expect(extractMatrixRange(engine, adr('A1', 1))).toEqual(new AbsoluteCellRange(adr('A1'), adr('B2')))
+  })
+
+  it('should be possible to remove row above array', () => {
+    const engine = HyperFormula.buildFromArray([
+      [],
+      ['=-C2:D4'],
+      [],
+      [],
+      ['foo']
+    ], {useArrayArithmetic: true})
+
+    engine.removeRows(0, [0, 1])
+
+    const expected = HyperFormula.buildFromArray([
+      ['=-C1:D3'],
+      [],
+      [],
+      ['foo']
+    ], {useArrayArithmetic: true})
+
+    expectEngineToBeTheSameAs(engine, expected)
+  })
+
+  it('removing row across array should not change array', () => {
+    const engine = HyperFormula.buildFromArray([
+      [1, 2], [3, 4], [5, 6],
+      ['=-A1:B3'],
+      [], [], [],
+      ['foo']
+    ], {useArrayArithmetic: true})
+
+    engine.removeRows(0, [4, 1])
+
+    expectEngineToBeTheSameAs(engine, HyperFormula.buildFromArray([
+      [1, 2], [3, 4], [5, 6],
+      ['=-A1:B3'],
+      [], [],
+      ['foo']
+    ], {useArrayArithmetic: true}))
+  })
+
+  it('removing row should shrink dependent array', () => {
+    const engine = HyperFormula.buildFromArray([
+      [1, 2],
+      [],
+      [3, 4],
+      ['=TRANSPOSE(A1:B3)']
+    ], {useArrayArithmetic: true})
+
+    engine.removeRows(0, [1, 1])
+
+    expectEngineToBeTheSameAs(engine, HyperFormula.buildFromArray([
+      [1, 2],
+      [3, 4],
+      ['=TRANSPOSE(A1:B2)']
+    ], {useArrayArithmetic: true}))
+  })
+
+  it('it should be REF if no space after removing row', () => {
+    const engine = HyperFormula.buildFromArray([
+      ['=-B3:B4'],
+      [],
+      [1, 1],
+      [null, 2],
+    ], {useArrayArithmetic: true})
+
+    engine.removeRows(0, [1, 1])
+
+    expect(engine.getSheetValues(0)).toEqual([
+      [noSpace()],
+      [1, 1],
+      [null, 2],
+    ])
+
+    const expected = HyperFormula.buildFromArray([
+      ['=-B2:B3'],
+      [1, 1],
+      [null, 2]
+    ], {useArrayArithmetic: true})
+    expectEngineToBeTheSameAs(engine, expected)
+  })
+
+  it('it should be REF, not CYCLE, after removing rows', () => {
+    const engine = HyperFormula.buildFromArray([
+      ['=-A3:A4'],
+      [],
+      [1],
+      [2]
+    ], {useArrayArithmetic: true})
+
+    engine.removeRows(0, [1, 1])
+
+    expect(engine.getSheetValues(0)).toEqual([
+      [noSpace()],
+      [1],
+      [2]
+    ])
+
+    const expected = HyperFormula.buildFromArray([
+      ['=-A2:A3'],
+      [1],
+      [2]
+    ], {useArrayArithmetic: true})
+    expectEngineToBeTheSameAs(engine, expected)
   })
 })
